working version of flask_sqlalchemy
